refactor: split builder types into dedicated file and fix WS browser signal double-meta bug
- Extract type definitions from builder.ts (608 lines) into builder-types.ts (332 lines), bringing builder.ts to 309 lines — both well under the 400-line limit.

- Fix signal() in transport-websocket-browser computing meta twice — remove redundant toPriorityMeta/toMessageMeta calls since mergeMeta already includes them.

Addresses audit findings C-1 (file size), M-4 (duplicate types), M-9 (double meta).
